+++ Upload/inc/plugins/feedpublisher.php
@@ -22,7 +22,7 @@ function feedpublisher_info()
         'website' => 'https://sickgaming.net',
         'author' => 'SickProdigy',
         'authorsite' => 'https://sickgaming.net',
-        'version' => '0.1.13',
+        'version' => '0.1.14',
         'compatibility' => '18*',
         'codename' => 'feedpublisher',
     );
@@ -50,6 +50,8 @@ function feedpublisher_install()
             `url` varchar(2048) NOT NULL,
             `fid` int unsigned NOT NULL,
             `uid` int unsigned NOT NULL,
+            `title_prefix` varchar(40) NOT NULL DEFAULT '',
+            `thread_prefix_id` int unsigned NOT NULL DEFAULT 0,
             `enabled` tinyint(1) NOT NULL DEFAULT 0,
             `interval_minutes` smallint unsigned NOT NULL DEFAULT 60,
             `publish_interval_minutes` smallint unsigned NOT NULL DEFAULT 60,
@@ -101,6 +103,8 @@ function feedpublisher_upgrade_schema()
     }
 
     $columns = array(
+        'title_prefix' => "varchar(40) NOT NULL DEFAULT '' AFTER `uid`",
+        'thread_prefix_id' => "int unsigned NOT NULL DEFAULT 0 AFTER `title_prefix`",
         'interval_minutes' => "smallint unsigned NOT NULL DEFAULT 60 AFTER `enabled`",
         'publish_interval_minutes' => "smallint unsigned NOT NULL DEFAULT 60 AFTER `interval_minutes`",
         'max_posts_per_run' => "smallint unsigned NOT NULL DEFAULT 1 AFTER `publish_interval_minutes`",